Add Optional STARTING_LED Macro
Setting this macro will push the start position of the controlled LEDs
down the strip. E.g., with a STARTING_LED of 10 & an LED_COUNT of 5,
LEDs number 10 through 14 of a strip will be controlled by the library.

We use this when we don't have enough amperage to power a whole strip &
want the patterns to be pushed to the end of the saber instead of
lighting up the LEDs near the hilt.