Add a Patterns Module for Playback of Generic Color Sequences
Add an `apa102_patterns` module. This module writes sequences to an
array which can then be output to the LEDs. This allows exension &
retraction of arbitrary patterns - unlike the Simple Effects module
which only supports extension & retraction of solid colors. The module
also supports specifying & playing back generic patterns.

Each pattern type(flash, scroll, etc.) has a `_step_count` &
`_set_sequence` function. The step count is the total number of discrete
states a pattern has. The _set_sequence functions write the LED color
sequence for the current step to the global array. Pattern playback is
done by initializing a pattern(setting the total steps & reseting the
current step), generating the sequence array, outputting the array,
delaying for an appropriate amount of time, & incrementing the current
step.

In the future this might be refactored so the sequence functions are
dependent on the LED index as well. This would allow writing different
patterns to different ranges of LEDs.